SimpleChart v1.2
Improved the way the chart's datasets are built. Now we choose the
average between 100 or 1000 samples, depending on the total number of
samples of the audio file. The samples chart works fine with datasets of
more than 1 million samples, but the response time is still high for
datasets over 10 million samples.

The energy sound chart works in the same way, but for datasets over 10
million samples doesn't work.